fix: normalize agent stream completion handling
